visualization/requirement.txt
matplotlib==3.7.2
streamlit==1.26.0
numpy==1.24.3
pillow==10.0.0
torch
xarray